Allow domains where TLD is >6 characters
It appears this library assumes the domain TLD will not be more than 6 characters but these days, many are: https://newgtlds.icann.org/en/program-status/delegated-strings

I'm not terribly well versed in RFC 1035, but in [Section 2.3.1](https://tools.ietf.org/html/rfc1035#section-2.3.1) there doesn't seem to be any mention about a maximum length so I'm curious about how this constraint ended up in here, and how necessary it is. I wonder if it might make sense to remove this constraint altogether.

Regardless, this change is trivial; it just allows for domains with a longer TLD. The length that I changed it to is fairly arbitrary. From the aforementioned list, it seems the longest (as of March 2019) is 18 characters. This change increases it to 26 characters; again, fairly arbitrarily.

I don't often open pull-requests, but when I do, it's for a one character change.